% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeEpileptogenicIndex}
\alias{computeEpileptogenicIndex}
\title{compute the Epileptogenic Index Object}
\usage{
computeEpileptogenicIndex(epoch, windowParams, fs = 1000)
}
\arguments{
\item{epoch}{Matrix or Epoch object. iEEG data matrix or Epoch object. If matrix, the row names are the electrode names and the column names are the time points}

\item{windowParams}{Numeric. Window parameters of the multitaper method}

\item{fs}{Numeric. frequency of signal iEEG acquisition}
}
\value{
A power analysis matrix
}
\description{
compute the Epileptogenic Index Object
}
\examples{
data("pt01EcoG")
windowParams<-c(0.25,0.1)
epoch <- Epoch(pt01EcoG)
EIPt01<-computeEpileptogenicIndex <- function(epoch, windowParams)
}
