% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_EpileptogenicIndex.r
\name{nrow,EpileptogenicIndex-method}
\alias{nrow,EpileptogenicIndex-method}
\alias{ncol,EpileptogenicIndex-method}
\title{Get the number of rows or columns of a EpileptogenicIndex object}
\usage{
\S4method{nrow}{EpileptogenicIndex}(x)

\S4method{ncol}{EpileptogenicIndex}(x)
}
\arguments{
\item{x}{A EpileptogenicIndex object}
}
\description{
Get the number of rows or columns of a EpileptogenicIndex object
}
