% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotERHeatmap}
\alias{plotERHeatmap}
\title{Visualization function ( energy ratio matrix)}
\usage{
plotERHeatmap(ei, sozIndex = NULL)
}
\arguments{
\item{sozIndex}{Integer or string. A group of electrodes to mark as in the Seizure Onset Zone (SOZ)}

\item{EI}{EpileptogenicIndex object from \code{computeEpileptogenicIndex}}
}
\value{
A ggplot object
}
\description{
\code{plotERHeatmap}: plot energy ratio heatmap with electrodes marked as soz colored
}
\examples{

data("pt01EcoG")

## sozIndex is the index of the electrodes we assume are in the SOZ
sozIndex <- attr(pt01EcoG, "sozIndex")
## precomputed Epileptogenic Index object
data("pt01EI")

## plot the mean power heatmap
plotER<-plotERHeatmap(ei=pt01EI,sozIndex=sozIndex)
plotER
}
