% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitaper.R
\name{process_spectrogram_params}
\alias{process_spectrogram_params}
\title{Process spectrogram inputs.
Helper function to create frequency vector and window indices}
\usage{
process_spectrogram_params(
  fs,
  nfft,
  frequency_range,
  window_start,
  datawin_size
)
}
\arguments{
\item{fs}{(numeric): sampling frequency in Hz  -- required}

\item{nfft}{(numeric): length of signal to calculate fft on -- required}

\item{frequency_range}{(numeric vector): c(\if{html}{\out{<min frequency>}}, \if{html}{\out{<max frequency>}}) (default: NULL, adjusted to
c(0, nyquist) later)}

\item{window_start}{(numeric vector): timestamps representing the beginning time for each window -- required}

\item{datawin_size}{datawin_size (numeric): seconds in one window -- required}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   window_idxs (matrix): indices of timestamps for each window (nxm where n=number of windows and m=datawin_size)
   stimes (numeric vector): times for the centers of the spectral bins (1xt)
   sfreqs (numeric vector): frequency bins for spectrogram (1xf)
   freq_inds (logical vector): indicates which frequencies are being analyzed in an array of frequencies from 0 to fs
               with steps of fs/nfft
}\if{html}{\out{</div>}}
}
\description{
Process spectrogram inputs.
Helper function to create frequency vector and window indices
}
\examples{
# Set up spectrogram parameters
res <- process_spectrogram_params(fs, nfft, frequency_range, window_start, winsize_samples)
window_idxs <- res[[1]]
stimes <- res[[2]]
sfreqs <- res[[3]]
freq_inds <- res[[4]]
}
