% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pt01Epoch}
\alias{pt01Epoch}
\title{Pt01 seizure 1 (-10:10s) around seizure onset}
\format{
pt01Epoch: A Matrix with 3000 rows (time points) and 84 columns (electrodes)
}
\source{
Fragility Multi-Center Retrospective Study
(\href{https://openneuro.org/datasets/ds003029/versions/1.0.0}{OpenNeuro})
}
\usage{
## EEG data
data(pt01Epoch)
}
\description{
This data corresponds to the first seizure of patient PT01 from
the Fragility Data Set.
The data contains only the good channels.
It has been notch filtered and common average referenced in RAVE.
The acquisition frequency is 1000 Hz
EcoG recording gathered in collaboration with the National Institute of Health
}
\keyword{datasets}
