% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_EpileptogenicIndex.r
\name{show,EpileptogenicIndex-method}
\alias{show,EpileptogenicIndex-method}
\title{Print the EpileptogenicIndex object}
\usage{
\S4method{show}{EpileptogenicIndex}(object)
}
\arguments{
\item{object}{A EpileptogenicIndex object}
}
\description{
Print the EpileptogenicIndex object
}
