% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_EpileptogenicIndex.r
\name{[,EpileptogenicIndex,ANY,ANY,ANY-method}
\alias{[,EpileptogenicIndex,ANY,ANY,ANY-method}
\title{Subset a EpileptogenicIndex object}
\usage{
\S4method{[}{EpileptogenicIndex,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A EpileptogenicIndex object}

\item{i}{A logical vector or a numeric vector of indices to subset the electrodes}

\item{j}{A logical vector or a numeric vector of indices to subset the time windows}

\item{...}{Additional arguments (not used)}

\item{drop}{Additional arguments (not used)}
}
\description{
Subset a EpileptogenicIndex object
}
