% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visuIEEGData}
\alias{visuIEEGData}
\title{Visualization of ictal iEEG}
\usage{
visuIEEGData(
  ieegts,
  timeRange = NULL,
  title = "Patient name seizure number",
  display = NULL
)
}
\arguments{
\item{ieegts}{Matrix or Fragility object. Either a matrix of iEEG time
series x(t), with time points as rows and electrodes names as columns,
or a Fragility object from \code{calcAdjFrag}}

\item{title}{String. Figure title}

\item{display}{Integer or string. Vector electrodes to display}
}
\value{
plot raw signal
}
\description{
Visualization of ictal iEEG
}
\examples{
data("pt01Epoch")
sozIndex <- attr(pt01Epoch,"sozIndex")
display <- c(sozIndex,77:80)
timeRange <- c(-10,10)
iEEGplot<-visuIEEGData(ieegts=pt01Epoch,timeRange=timeRange,display=display)
iEEGplot
}
